workspaces/backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.4,<12.0
pytesseract>=0.3.10,<1.0
